services/api/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0
dj-database-url==2.1.0

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
cryptography==41.0.7

# Background Tasks
celery==5.3.4
celery[redis]==5.3.4
redis==4.6.0

# Message Queue
pika==1.3.2

# Search
elasticsearch==8.10.1
elasticsearch-dsl==8.11.0

# NLP and Classification
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4

# Monitoring & Observability
prometheus-client==0.19.0
django-prometheus==2.3.1
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.2.2
responses==0.24.1

# WhatsApp Integration & HTTP
requests==2.31.0
httpx==0.25.2

# Utilities
python-decouple==3.8
Pillow==10.1.0
python-magic==0.4.27
validators==0.22.0

# Development & Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
django-stubs==4.2.7
pre-commit==3.5.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0